fix: RBAC apiGroup, PDB handling, type safety, conflicts_with
- Fix eviction RBAC: use core API group, remove unnecessary delete permission
- Handle PDB rejections gracefully (429/500 are expected, not errors)
- Change crash_grace_period to u32 (reject negatives at CLI boundary)
- Add conflicts_with between --profile and --faults
- Add Latency to property-based scheduler tests (pre-existing gap)

+++ crates/cli/src/main.rs
@@ -141,10 +141,10 @@ struct RunArgs {
     /// Grace period in seconds for crash faults (default: K8s default 30s).
     /// Use 0 for immediate kill (no SIGTERM).
     #[arg(long)]
-    crash_grace_period: Option<i64>,
+    crash_grace_period: Option<u32>,
 
     /// Fault profile preset. Overrides --faults when set.
-    #[arg(long, value_enum)]
+    #[arg(long, value_enum, conflicts_with = "faults")]
     profile: Option<FaultProfile>,
 }
 
@@ -300,10 +300,10 @@ struct ExploreArgs {
     /// Grace period in seconds for crash faults (default: K8s default 30s).
     /// Use 0 for immediate kill (no SIGTERM).
     #[arg(long)]
-    crash_grace_period: Option<i64>,
+    crash_grace_period: Option<u32>,
 
     /// Fault profile preset. Overrides --faults when set.
-    #[arg(long, value_enum)]
+    #[arg(long, value_enum, conflicts_with = "faults")]
     profile: Option<FaultProfile>,
 }
 
@@ -1154,7 +1154,7 @@ async fn run_single_simulation(
     warmup: std::time::Duration,
     faults: &[String],
     inject_method: InjectMethod,
-    crash_grace_period: Option<i64>,
+    crash_grace_period: Option<u32>,
     property_defs: &[properties::PropertyDef],
 ) -> Result<SimulationResult> {
     let handle = heisensim_timeline::TimelineHandle::new();
